# Add after_preamble hook for precompiled_with_ambles and separate Rails-related code