--- a/lib/haml/temple_engine.rb
+++ b/lib/haml/temple_engine.rb
@@ -58,13 +58,12 @@ module Haml
     # (see {file:REFERENCE.md#encodings the `:encoding` option}).
     #
     # @return [String]
-    def precompiled_with_ambles(local_names)
-      preamble = <<END.tr!("\n", ';')
+    def precompiled_with_ambles(local_names, after_preamble: '')
+      preamble = "#{<<END};#{after_preamble};".tr!("\n", ';')
 begin
 extend Haml::Helpers
 _hamlout = @haml_buffer = Haml::Buffer.new(haml_buffer, #{Options.new(options).for_buffer.inspect})
 _erbout = _hamlout.buffer
-@output_buffer = output_buffer ||= ActionView::OutputBuffer.new rescue nil
 END
       postamble = <<END.tr!("\n", ';')
 #{precompiled_method_return_value}