--- a/lib/haml/temple_engine.rb
+++ b/lib/haml/temple_engine.rb
@@ -59,11 +59,12 @@ module Haml
     #
     # @return [String]
     def precompiled_with_ambles(local_names, after_preamble: '')
-      preamble = "#{<<END};#{after_preamble};".tr!("\n", ';')
+      preamble = <<END.tr!("\n", ';')
 begin
 extend Haml::Helpers
 _hamlout = @haml_buffer = Haml::Buffer.new(haml_buffer, #{Options.new(options).for_buffer.inspect})
 _erbout = _hamlout.buffer
+#{after_preamble}
 END
       postamble = <<END.tr!("\n", ';')
 #{precompiled_method_return_value}