@k0kubun 2017-02-06
While I'm sure we should have this option, I'm not 100% confident about the name `after_preamble`. Please give me some opinions.